Fixed incorrect inversion of policy as rules conditions #3419
