+++ tests/PSRule.Rules.Azure.Tests/Policy.assignment.10.json
@@ -1,0 +1,115 @@
+[
+  {
+    "Identity": null,
+    "Location": null,
+    "Name": "assignment.10",
+    "ResourceId": "/providers/Microsoft.Management/managementGroups/00000000-0000-0000-0000-000000000000/providers/Microsoft.Authorization/policyAssignments/assignment.10",
+    "ResourceName": "assignment.10",
+    "ResourceGroupName": null,
+    "ResourceType": "Microsoft.Authorization/policyAssignments",
+    "SubscriptionId": null,
+    "Sku": null,
+    "PolicyAssignmentId": "/providers/Microsoft.Management/managementGroups/00000000-0000-0000-0000-000000000000/providers/Microsoft.Authorization/policyAssignments/assignment.10",
+    "Properties": {
+      "Scope": "/providers/Microsoft.Management/managementGroups/00000000-0000-0000-0000-000000000000",
+      "NotScopes": null,
+      "DisplayName": "Org Audit Policies",
+      "Description": "https://docs.contoso.com/azure/policies",
+      "Metadata": {
+        "Version": "20240706"
+      },
+      "EnforcementMode": 0,
+      "PolicyDefinitionId": "/providers/Microsoft.Management/managementGroups/00000000-0000-0000-0000-000000000000/providers/Microsoft.Authorization/policySetDefinitions/assignment.10",
+      "Parameters": null,
+      "NonComplianceMessages": null
+    },
+    "PolicyDefinitions": [
+      {
+        "Name": "policyDefinition.1",
+        "ResourceId": "/providers/Microsoft.Authorization/policyDefinitions/policyDefinition.1",
+        "ResourceName": "policyDefinition.1",
+        "ResourceType": "Microsoft.Authorization/policyDefinitions",
+        "SubscriptionId": null,
+        "Properties": {
+          "Description": "Test case for https://github.com/Azure/PSRule.Rules.Azure/issues/3419.",
+          "DisplayName": "Org tagging",
+          "Metadata": {
+            "version": "1.0.0",
+            "category": "Security Center"
+          },
+          "Mode": "All",
+          "Parameters": {
+            "effect": {
+              "type": "String",
+              "metadata": {
+                "displayName": "Effect",
+                "description": "Enable or disable the execution of the policy"
+              },
+              "allowedValues": [
+                "Audit",
+                "Deny",
+                "Disabled"
+              ],
+              "defaultValue": "Audit"
+            },
+            "environments": {
+              "type": "Array",
+              "allowedValues": [
+                "Development",
+                "Staging",
+                "Production"
+              ],
+              "defaultValue": [
+                "Development"
+              ]
+            }
+          },
+          "PolicyRule": {
+            "if": {
+              "allOf": [
+                {
+                  "field": "type",
+                  "in": [
+                    "Microsoft.Sql/servers/databases",
+                    "Microsoft.Sql/servers",
+                    "Microsoft.KeyVault/vaults",
+                    "Microsoft.Cache/redis",
+                    "Microsoft.Storage/storageAccounts",
+                    "Microsoft.ServiceBus/namespaces"
+                  ]
+                },
+                {
+                  "field": "tags.Environment",
+                  "in": "[parameters('environments')]"
+                },
+                {
+                  "not": {
+                    "field": "tags[NSW]",
+                    "in": [
+                      "0",
+                      "10",
+                      "20",
+                      "30",
+                      "40",
+                      "50"
+                    ]
+                  }
+                }
+              ]
+            },
+            "then": {
+              "effect": "[parameters('effect')]"
+            }
+          },
+          "versions": [
+            "1.0.0"
+          ],
+          "PolicyType": 2
+        },
+        "PolicyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/policyDefinition.1",
+        "id": "/providers/Microsoft.Authorization/policyDefinitions/policyDefinition.1"
+      }
+    ],
+    "exemptions": []
+  }
+]
